pyproject.toml
[tool.poetry]
name = "altpe-sdk"
version = "0.1.0"
description = "Python SDK for Alternatives.PE VentureCap API"
authors = ["Brightriver"]
license = "MIT"
readme = "README.md"
packages = [{ include = "altpe_sdk" }]

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.25.2"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
ruff = "^0.1.6"
ipykernel = "^6.30.1"
rich = "^14.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
